+++ packages/cli/package.json
@@ -59,6 +59,8 @@
     "@octokit/oauth-app": "^4.2.0",
     "@octokit/request": "^6.0.0",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
+    "@pnpm/lockfile-file": "^9.1.3",
+    "@pnpm/lockfile-types": "^7.1.3",
     "@rollup/plugin-commonjs": "^26.0.0",
     "@rollup/plugin-json": "^6.0.0",
     "@rollup/plugin-node-resolve": "^15.0.0",
