+++ packages/cli/src/lib/versioning/packages.ts
@@ -16,9 +16,6 @@
 
 import { minimatch } from 'minimatch';
 import { getPackages } from '@manypkg/get-packages';
-import { NotFoundError } from '../errors';
-import { detectYarnVersion } from '../yarn';
-import { execFile } from '../run';
 
 const DEP_TYPES = [
   'dependencies',
@@ -27,69 +24,11 @@ const DEP_TYPES = [
   'optionalDependencies',
 ] as const;
 
-// Package data as returned by `yarn info`
-export type YarnInfoInspectData = {
-  name: string;
-  'dist-tags': Record<string, string>;
-  versions: string[];
-  time: { [version: string]: string };
-};
-
-// Possible `yarn info` output
-type YarnInfo = {
-  type: 'inspect';
-  data: YarnInfoInspectData | { type: string; data: unknown };
-};
-
 type PkgVersionInfo = {
   range: string;
   name: string;
   location: string;
 };
-
-export async function fetchPackageInfo(
-  name: string,
-): Promise<YarnInfoInspectData> {
-  const yarnVersion = await detectYarnVersion();
-
-  const cmd = yarnVersion === 'classic' ? ['info'] : ['npm', 'info'];
-  try {
-    const { stdout: output } = await execFile(
-      'yarn',
-      [...cmd, '--json', name],
-      { shell: true },
-    );
-
-    if (!output) {
-      throw new NotFoundError(
-        `No package information found for package ${name}`,
-      );
-    }
-
-    if (yarnVersion === 'berry') {
-      return JSON.parse(output) as YarnInfoInspectData;
-    }
-
-    const info = JSON.parse(output) as YarnInfo;
-    if (info.type !== 'inspect') {
-      throw new Error(`Received unknown yarn info for ${name}, ${output}`);
-    }
-
-    return info.data as YarnInfoInspectData;
-  } catch (error) {
-    if (yarnVersion === 'classic') {
-      throw error;
-    }
-
-    if (error?.stdout.includes('Response Code: 404')) {
-      throw new NotFoundError(
-        `No package information found for package ${name}`,
-      );
-    }
-
-    throw error;
-  }
-}
 
 /** Map all dependencies in the repo as dependency => dependents */
 export async function mapDependencies(
